pyproject.toml
[tool.poetry]
name = "fif_recsys"
version = "0.0.1"
description = "Ranking Brazilian fixed income funds"
authors = ["gustavo.polleti <gustavo.polleti@gmail.com>"]
readme = "README.md"
packages = [{include = "fif_recsys"}]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
]

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.17.0"
rich = "^14.0.0"
shellingham = "^1.5.0"
requests = "^2.32.0"
# Core data dependencies required at runtime
pandas = "^2.1.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
tox = "*"
flake8 = "*"
black = "*"
mypy = "*"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.6.0"
mkdocs-markdownextradata-plugin = "^0.2.6"
# Needed to convert notebooks to markdown for inclusion in the MkDocs site
jupyter = "*"
nbconvert = "*"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.0"

[tool.poetry.group.publish.dependencies]
twine = "^6.0.0"

[tool.poetry.scripts]
fif = "fif_recsys.main:fif"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
